+++ src/aux-common/partitions/RemoteCausalRepoHistoryPartition.ts
@@ -175,7 +175,7 @@ export class RemoteCausalRepoHistoryPartitionImpl
                 }
                 const hash = bot.tags.markHash;
                 this._client
-                    .restore(restoreMark.server || this._branch, hash)
+                    .restore(restoreMark.inst || this._branch, hash)
                     .subscribe(
                         () => {
                             if (hasValue(event.taskId)) {